requirements.txt
# ABC Music Manager — runtime dependencies
# Python 3.10+ recommended
# UI: PySide6 (Qt for Python) — see DECISIONS 002, 026

PySide6>=6.6.0
superqt
Send2Trash>=1.8.0

# Playback (ABC to MIDI, FluidSynth)
pyfluidsynth>=1.3.0
mido>=1.3.0
